requirements.txt
streamlit
openai
load_dotenv
PyPDF2
python-docx
python-pptx
azure-cosmos>=4.5.1
azure-identity>=1.15.0
reportlab
azure-storage-blob>=12.19.0
numpy>=1.24.0
tavily-python>=0.3.0
# Optional: for HuggingFace zero-shot classification
# transformers>=4.30.0
# torch>=2.0.0